Hide the toolbar's center element (quick search) if we don't display it
This allows us to get rid of the gap between the toolbars that shows up
as "unstyled" background in some themes like Greybird.

I couldn't find a way to get the background color to display properly if
the quick search is displayed, I fear this might be a theme issue.

Experiments with the `primary-toolbar` style class shows mixed results,
so I dropped most of those changes.